+++ .env.example
@@ -34,5 +34,5 @@ DATABASE_URL="postgresql://postgres:password@localhost:5432/schroedinger-hat-web
 # Secret for securing cron job endpoints (generate a secure random string)
 CRON_SECRET="your-secure-random-string"
 
-# Postmark API key
-POSTMARK_API_KEY="your-postmark-api-key-here"
+# Email
+RESEND_API_KEY="your-resend-api-key-here"
