+++ package.json
@@ -64,13 +64,13 @@
     "motion": "^11.15.0",
     "next": "^15.1.0",
     "next-sanity": "^9.8.27",
-    "postmark": "^4.0.5",
     "react": "^18.3.1",
     "react-confetti": "^6.1.0",
     "react-dom": "^18.3.1",
     "react-email": "^3.0.4",
     "react-hook-form": "^7.53.2",
     "react-syntax-highlighter": "^15.6.1",
+    "resend": "^4.1.1",
     "sanity": "^3.67.1",
     "server-only": "^0.0.1",
     "stripe": "^17.3.1",
