+++ src/server/email.tsx
@@ -1,0 +1,66 @@
+import { Resend } from "resend"
+import { render } from "@react-email/render"
+import { MembershipSignupEmail } from "@/emails/membership-signup"
+import { env } from "@/env"
+import React from "react"
+
+let _resend: any = null
+
+function resendIsConfigured() {
+  return _resend !== null
+}
+function configureResend() {
+  if (!env.RESEND_API_KEY) {
+    throw new Error("RESEND_API_KEY not set")
+  }
+
+  _resend = new Resend(env.RESEND_API_KEY as unknown as string)
+  return _resend
+}
+
+interface EmailPayload {
+  from?: string
+  to: string
+  subject: string
+  html: string
+}
+
+async function sendEmail(payload: EmailPayload) {
+  const { from = "hello@schroedinger-hat.org", to, subject, html } = payload
+
+  if (!resendIsConfigured()) {
+    console.log("RESEND_API_KEY not set. Email would have been sent with the following details:")
+    console.log({ from, to, subject, html })
+    return
+  }
+
+  const resend = configureResend()
+
+  try {
+    const { data, error } = await resend.emails.send({
+      from,
+      to,
+      subject,
+      html,
+    })
+
+    if (error) {
+      console.error("Failed to send email:", { error })
+    }
+
+    return data
+  } catch (error) {
+    console.error("Failed to send email:", { error })
+    throw error
+  }
+}
+
+export async function sendMembershipSignupEmail(firstName: string, email: string) {
+  const emailHtml = await render(<MembershipSignupEmail firstName={firstName} />, { pretty: true })
+
+  return sendEmail({
+    to: email,
+    subject: "Welcome to Schrödinger Hat!",
+    html: emailHtml,
+  })
+}
